timesheet_tool/requirements.txt
fastapi==0.98.0
pydantic<2.0.0
uvicorn==0.22.0
sqlalchemy==2.0.20
asyncpg==0.27.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.12.0
pytest
pytest-asyncio
httpx
pydantic[dotenv]
python-dotenv
starlette>=0.22.0
psycopg2-binary
 